Move components into src/app
