+++ pages/open-source.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import Page from '../components/Page'
-import Projects from '../components/Projects'
+import Page from '../src/app/components/Page'
+import Projects from '../src/app/components/Projects'
 import withCommonProps from '../src/app/propMiddleware/withCommonProps'
 
 export default function OpenSource ({ githubUrl, projects }) {
