+++ pages/thoughts/[slug].js
@@ -1,8 +1,8 @@
 import React from 'react'
 import Link from 'next/link'
-import Page from '../../components/Page'
-import Thought from '../../components/Thought'
-import RelatedContent from '../../components/RelatedContent'
+import Page from '../../src/app/components/Page'
+import Thought from '../../src/app/components/Thought'
+import RelatedContent from '../../src/app/components/RelatedContent'
 import useDelayedRelatedContent from '../../src/app/hooks/useDelayedRelatedContent'
 import useLiveBlog from '../../src/app/hooks/useLiveBlog'
 import { fetchOneThoughtBySlug } from '../../src/app/factory'
