+++ pages/topics/[slug].js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Link from 'next/link'
-import Page from '../../components/Page'
-import Thoughts from '../../components/Thoughts'
+import Page from '../../src/app/components/Page'
+import Thoughts from '../../src/app/components/Thoughts'
 import { fetchThoughtsByTopicSlug, fetchTopicBySlug } from '../../src/app/factory'
 import withErrorHandling from '../../src/app/propMiddleware/withErrorHandling'
 import withCommonProps from '../../src/app/propMiddleware/withCommonProps'
