+++ src/app/components/Thought/index.js
