+++ src/app/components/Thoughts/index.js
