+++ test/pages/about.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import About from '../../pages/about'
-import Page from '../../components/Page'
+import Page from '../../src/app/components/Page'
 
 test('content renders', () => {
   const page = shallow(<About />)
