+++ test/pages/index.test.js
@@ -2,8 +2,8 @@ import React from 'react'
 import { shallow } from 'enzyme'
 import nock from 'nock'
 import Index from '../../pages/index'
-import Page from '../../components/Page'
-import Thoughts from '../../components/Thoughts'
+import Page from '../../src/app/components/Page'
+import Thoughts from '../../src/app/components/Thoughts'
 
 test('content renders', () => {
   const page = shallow(<Index />)
